refactor(number-theory): isolate powerful contracts
