+++ src/jacobian/math/number_theory/_powerful.py
@@ -4,13 +4,13 @@ from __future__ import annotations
 
 from jacobian.canonical import format_canonical_integer, parse_canonical_integer
 from jacobian.catalog._examples import example
-from jacobian.math.number_theory._models import (
+from jacobian.math.number_theory._models import PrimePower
+from jacobian.math.number_theory._powerful_kernels import decide_powerful_data
+from jacobian.math.number_theory._powerful_models import (
     PowerfulNumberRequest,
     PowerfulNumberResult,
-    PrimePower,
     ResidualPerfectPower,
 )
-from jacobian.math.number_theory._powerful_kernels import decide_powerful_data
 from jacobian.math.number_theory._support import number_theory_operation
